requirements.txt
streamlit==1.31.1
django==4.2.9
djangorestframework==3.14.0
chromadb==0.4.22
langchain==0.1.9
langchain-groq==0.0.1
python-dotenv==1.0.1
huggingface-hub==0.20.3
sentence-transformers==2.5.1
django-cors-headers==4.3.1 